requirements.txt
sentence-transformers 
psycopg2-binary 
numpy 
tqdm 
pandas 